+++ desktop/src/jvmMain/kotlin/ui/components/outlined_radio_button/internal/CenteringBox.kt
@@ -1,0 +1,19 @@
+@file:Suppress("FunctionName")
+
+package ui.components.outlined_radio_button.internal
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+internal fun CenteringBox(content: @Composable () -> Unit) {
+    Box(
+        modifier = Modifier.fillMaxWidth(),
+        contentAlignment = Alignment.Center
+    ) {
+        content.invoke()
+    }
+}
